requirements.txt
tensorflow
numpy
opencv-python
moviepy
pillow